Bug Fixes
git-svn-id: https://opensynx.svn.codeplex.com/svn@9606 937343c3-697a-4e39-b5c6-83b50a1b7d63

+++ Synx/Synx/Form1.vb
@@ -185,19 +185,39 @@ Public Class Home
 
     Private Sub Button9_Click(ByVal sender As System.Object, ByVal e As System.EventArgs) Handles Button9.Click
         'This is just used for working with the AirSynx FTP Upload Tool.
-        If My.Computer.FileSystem.FileExists("C:\ProgramData\Synx\asxftpuV1.0.0.0\asxftpu.exe") Then
-            System.IO.File.WriteAllText("C:\ProgramData\Synx\asxun.dat", username.Text)
-            System.IO.File.WriteAllText("C:\ProgramData\Synx\asxpw.dat", password.Text)
-            Process.Start("C:\ProgramData\Synx\asxftpu.exe")
+        If My.Computer.FileSystem.DirectoryExists("C:\ProgramData\Synx") Then
+            If My.Computer.FileSystem.FileExists("C:\ProgramData\Synx\asxftpuV1.0.0.0\asxftpu.exe") Then
+                System.IO.File.WriteAllText("C:\ProgramData\Synx\asxun.dat", username.Text)
+                System.IO.File.WriteAllText("C:\ProgramData\Synx\asxpw.dat", password.Text)
+                Process.Start("C:\ProgramData\Synx\asxftpuV1.0.0.0\asxftpu.exe")
+            Else
+                My.Computer.FileSystem.CreateDirectory("C:\ProgramData\Synx\asxftpuV1.0.0.0")
+                MessageBox.Show("Starting Download...")
+                My.Computer.Network.DownloadFile("http://content.etheralstudios.com/AirSynx/asxftpu.exe", "C:\ProgramData\Synx\asxftpuV1.0.0.0\asxftpu.exe")
+                My.Computer.Network.DownloadFile("http://content.etheralstudios.com/AirSynx/asxftpu.vshost.exe.manifest", "C:\ProgramData\Synx\asxftpuV1.0.0.0\asxftpu.vshost.exe.manifest")
+                My.Computer.Network.DownloadFile("http://content.etheralstudios.com/AirSynx/asxftpu.vshost.exe", "C:\ProgramData\Synx\asxftpuV1.0.0.0\asxftpu.vshost.exe")
+                MessageBox.Show("Done!")
+                System.IO.File.WriteAllText("C:\ProgramData\Synx\asxun.dat", username.Text)
+                System.IO.File.WriteAllText("C:\ProgramData\Synx\asxpw.dat", password.Text)
+                Process.Start("C:\ProgramData\Synx\asxftpuV1.0.0.0\asxftpu.exe")
+            End If
         Else
-            MessageBox.Show("Starting Download...")
-            My.Computer.Network.DownloadFile("http://content.etheralstudios.com/AirSynx/asxftpu.exe", "C:\ProgramData\Synx\asxftpuV1.0.0.0")
-            My.Computer.Network.DownloadFile("http://content.etheralstudios.com/AirSynx/asxftpu.vshost.exe.manifest", "C:\ProgramData\Synx\asxftpuV1.0.0.0")
-            My.Computer.Network.DownloadFile("http://content.etheralstudios.com/AirSynx/asxftpu.vshost.exe", "C:\ProgramData\Synx\asxftpuV1.0.0.0")
-            MessageBox.Show("Done!")
-            System.IO.File.WriteAllText("C:\ProgramData\Synx\asxun.dat", username.Text)
-            System.IO.File.WriteAllText("C:\ProgramData\Synx\asxpw.dat", password.Text)
-            Process.Start("C:\ProgramData\Synx\asxftpu.exe")
+            My.Computer.FileSystem.CreateDirectory("C:\ProgramData\Synx")
+            If My.Computer.FileSystem.FileExists("C:\ProgramData\Synx\asxftpuV1.0.0.0\asxftpu.exe") Then
+                System.IO.File.WriteAllText("C:\ProgramData\Synx\asxun.dat", username.Text)
+                System.IO.File.WriteAllText("C:\ProgramData\Synx\asxpw.dat", password.Text)
+                Process.Start("C:\ProgramData\Synx\asxftpuV1.0.0.0\asxftpu.exe")
+            Else
+                My.Computer.FileSystem.CreateDirectory("C:\ProgramData\Synx\asxftpuV1.0.0.0")
+                MessageBox.Show("Starting Download...")
+                My.Computer.Network.DownloadFile("http://content.etheralstudios.com/AirSynx/asxftpu.exe", "C:\ProgramData\Synx\asxftpuV1.0.0.0\asxftpu.exe")
+                My.Computer.Network.DownloadFile("http://content.etheralstudios.com/AirSynx/asxftpu.vshost.exe.manifest", "C:\ProgramData\Synx\asxftpuV1.0.0.0\asxftpu.vshost.exe.manifest")
+                My.Computer.Network.DownloadFile("http://content.etheralstudios.com/AirSynx/asxftpu.vshost.exe", "C:\ProgramData\Synx\asxftpuV1.0.0.0\asxftpu.vshost.exe")
+                MessageBox.Show("Done!")
+                System.IO.File.WriteAllText("C:\ProgramData\Synx\asxun.dat", username.Text)
+                System.IO.File.WriteAllText("C:\ProgramData\Synx\asxpw.dat", password.Text)
+                Process.Start("C:\ProgramData\Synx\asxftpuV1.0.0.0\asxftpu.exe")
+            End If
         End If
     End Sub
 End Class
